third_party/WebKit/LayoutTests/external/wpt/css/css-tables/fixed-layout-2-expected.txt

This is a testharness.js-based test.
PASS Table-layout:fixed is not applied when width is auto
FAIL Table-layout:fixed reports it is not applied when width is auto assert_equals: expected "auto" but got "fixed"
FAIL Table-layout:fixed is not applied when width is max-content assert_equals: expected 100 but got 0
FAIL Table-layout:fixed reports it is not applied when width is max-content assert_equals: expected "auto" but got "fixed"
FAIL Table-layout:fixed is not applied when width is min-content assert_equals: expected 0 but got 100
Harness: the test ran to completion.
